+++ src/journal/journal.rs
@@ -70,10 +70,10 @@ impl<'a> UsnJournal<'a> {
         Ok(UsnJournalIter::new(
             self.volume.handle,
             journal_data.journal_id,
-            vec![0u8; options.buffer_size],
+            vec![0u8; options.buffer_bytes.get()],
             UsnJournalIterConfig {
                 next_start_usn: options.start_usn.get(),
-                reason_mask: options.reason_mask,
+                reason_mask: options.reason_mask.bits(),
                 return_only_on_close: options.only_on_close as u32,
                 timeout: options.timeout,
                 bytes_to_wait_for: options.wait_for_more as u64,